requirements-dev.txt
# Development dependencies
-r requirements.txt

# Code formatting and linting
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
flake8-django>=1.1.5
flake8-isort>=6.0.0

# Type checking
mypy>=1.0.0
django-stubs>=4.2.0
types-requests
types-python-dateutil

# Testing
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
factory-boy>=3.2.0

# Pre-commit hooks
pre-commit>=3.0.0